image/requirements.txt
numpy
Pillow
pydaisi
streamlit